Added E. E. Cummings factorial war poem
It's a little known fact that Cummings worked as one of the earliest
"embedded" reporters during World War I, and this poem appears to be
a brief rumination on his experiences in the war and his attempts to
remain unbiased amidst the chaos. As a side-effect, it computes
factorials, and also gets in some sly digs on the vagaries of
JavaScript punctuation.